chatbot-backend/requirements.txt
fastapi==0.115.6
uvicorn==0.25.0
python-dotenv==1.0.1
qdrant-client==1.12.1
google-genai==0.2.1
psycopg[binary]==3.2.3
SQLAlchemy==2.0.36
asyncpg==0.30.0
tiktoken==0.8.0
chainlit==1.3.2
httpx==0.28.1
alembic==1.13.1
